Create Preprocessed_Dict.py
Dictionary that contains information for each file included in the analysis.  ICA removed is the number of components removed during preprocessing and is used for rank reduction in the full script.  Stim channel and trigger value contain the information to generate the appropriate epochs, and the transform is the path to the transformation file made from the Save_Preprocessed_Data script.